cleanup from pr review
