+++ airbyte-webapp/src/views/Connection/ConnectionForm/components/EditControls.module.scss
@@ -15,21 +15,6 @@
   margin-left: variables.$spacing-md;
 }
 
-.message {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  flex: 1;
-}
-
-.success {
-  color: colors.$green;
-}
-
-.error {
-  color: colors.$red;
-}
-
 // currently only implemented on transformation view form card, margins are specific to that implementation
 // todo: standardize the margin sizes here
 .line {
