+++ airbyte-webapp/src/views/Connection/ConnectionForm/components/ResponseMessage.module.scss
@@ -1,0 +1,14 @@
+.message {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  flex: 1;
+}
+
+.success {
+  color: colors.$green;
+}
+
+.error {
+  color: colors.$red;
+}
